// packages/logging/src/node/enums.ts
// @generated by drool.exe extraction
// Generated from packages/logging/src/node/enums; source module was missing in Bun sourcemap recovery.
export enum RotationSyscall {
  Readdir = "readdir", // medium-cluster; compactCamel; bundle hits=7
  Rename = "rename", // medium-cluster; compactCamel; bundle hits=36
  Rm = "rm", // medium-cluster; compactCamel; bundle hits=39
  Stat = "stat", // medium-cluster; compactCamel; bundle hits=17
}